+++ integrations/amqp/v3/CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [v3.2.0] - 2026-07-02 - Message-Id Round-Trip and Bucketed Delay Topology
+## [v3.2.0] - 2026-07-03 - Message-Id Round-Trip and Bucketed Delay Topology
 
 ### Added